Updates to not loose progress
This is not a fully working version yet, of where I want it to be, haven't even tested it, but I'm sure it'll do fine after I work out the kinks and diffs between the definition file .d.ts and perhaps get that in as a real typedef, I plan on pushing this up to npm as well, potentially bower, nuget, whereever I can, but as of now it is current state, not tested, verified, and finished, docs, etc. Could give someone ideas in the mean time.