Building and starting to run, now to fix stackoverflow lol
Building and starting to run, now to fix stackoverflow lol. the Convert function is probably newing up things which in turn calls Convert again, need to redesign, I've been back and forth on it but now that its compiling and running from a single output file in the root directory/Index.html example this is so close to being tidied up minus I abondoned IArrayConstructor.ts temporarily, as it was a nightmare for not being able to matchup constructors for some reason, but I may add it back later just because, or I might leave it out comletely if it is a bane for my existence.